Add documentation generation with Doxygen
I had to modify Doxygen output directories because with meson we can't assume an execution directory and Doxygen doesn't create
intermediate subdirectories (eg: html/api). I have to assume that the other build systems were creating the subdirectories in
advance. The other build systems will need to be updated to match this configuration.